Moved regulars and notice of departure checkers to their own respective handler class
