+++ src/main/kotlin/dev/vxrp/bot/noticeofdeparture/NoticeOfDepartureManager.kt
@@ -1,5 +1,6 @@
 package dev.vxrp.bot.noticeofdeparture
 
+import dev.vxrp.bot.noticeofdeparture.handler.NoticeOfDepartureCheckerHandler
 import dev.vxrp.bot.noticeofdeparture.handler.NoticeOfDepartureMessageHandler
 import dev.vxrp.configuration.data.Config
 import dev.vxrp.configuration.data.Translation
@@ -10,8 +11,6 @@ import dev.vxrp.util.duration.DurationParser
 import dev.vxrp.util.duration.enums.DurationType
 import net.dv8tion.jda.api.JDA
 import org.slf4j.LoggerFactory
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 
 class NoticeOfDepartureManager(val api: JDA, val config: Config, val translation: Translation) {
     private val logger = LoggerFactory.getLogger(NoticeOfDepartureManager::class.java)
@@ -33,32 +32,6 @@ class NoticeOfDepartureManager(val api: JDA, val config: Config, val translation
             config.settings.noticeOfDeparture.checkRate,
             DurationType.valueOf(config.settings.noticeOfDeparture.checkUnit)),
             noticeOfDepartureScope
-        ) { checkerTask() }
-    }
-
-    private suspend fun checkerTask() {
-        val notices = NoticeOfDepartureTable().retrieveSerial()
-        logger.info("Starting notice of departure checker, processing units starting...")
-        val idList = NoticeOfDepartureTable().retrieveAllIds()
-        var redundantNotices = 0
-
-        for (id in idList) {
-            val currentDate = LocalDate.now()
-            val formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy")
-            val beginDate = LocalDate.parse(NoticeOfDepartureTable().retrieveBeginDate(id)!!, formatter)
-            val endDate = LocalDate.parse(NoticeOfDepartureTable().retrieveEndDate(id)!!, formatter)
-
-            val daysUntil = currentDate.until(endDate).days
-
-            if (daysUntil == 0 || daysUntil < 0) {
-                redundantNotices += 1
-                logger.info("Found redundant notice of departure, ending notice and processing database data")
-                NoticeOfDepartureMessageHandler(api, config, translation).sendEndedMessage(id, beginDate.format(formatter), endDate.format(formatter))
-                NoticeOfDepartureTable().deleteEntry(id)
-                logger.info("Deleted notice of departure - $id successfully")
-            }
-        }
-
-        logger.info("$redundantNotices out of $notices notices were found to be redundant, checker ending process...")
+        ) { NoticeOfDepartureCheckerHandler(api, config, translation).checkerTask() }
     }
 }
